Extract terminal service
